Share createOnSecret across harnesses.
